% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{convert_file}
\alias{convert_file}
\title{convert_file}
\usage{
convert_file(input, output, type = ropencc::S2T, inputEncoding = 'UTF-8', ...)
}
\arguments{
\item{input}{input file name}

\item{output}{output file name}

\item{type}{one of T2S S2T S2TW HK2S S2HK S2TWP T2HK T2S T2TW TW2S TW2SP}

\item{inputEncoding}{encoding for input}

\item{...}{other parameter pass to readLines function}
}

