% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hello.R
\name{converter}
\alias{converter}
\title{Converter}
\usage{
converter(type = ropencc::S2T)
}
\arguments{
\item{type}{T2S S2T S2TW HK2S S2HK S2TWP T2HK T2S T2TW TW2S TW2SP}
}
\description{
create a converter. S2T means Simplified Chinese to Traditional Chinese. T2S means Traditional Chinese to Simplified Chinese.
}
\examples{
\dontrun{
cc = converter(S2T)
}
}

