% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dict.R
\name{make_dict}
\alias{make_dict}
\title{Dict Convertion}
\usage{
make_dict(input = NULL, output = NULL, from_format = "text",
  to_format = "ocd")
}
\arguments{
\item{input}{input file}

\item{output}{output file}

\item{from_format}{from text or ocd}

\item{to_format}{to text or ocd}
}
\description{
Create dict from text file or create text file from ocd file.
}

