% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hello.R
\name{run_convert}
\alias{[.cc_converter}
\alias{run_convert}
\title{Run Converter}
\usage{
run_convert(worker, text)

\method{[}{cc_converter}(ptr, code)
}
\arguments{
\item{worker}{opencc worker}

\item{text}{input text}

\item{ptr}{worker}

\item{code}{input text}
}
\description{
run converter
}
\examples{
\dontrun{
cc = converter()
}
}

