% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{S2T}
\alias{HK2S}
\alias{S2HK}
\alias{S2T}
\alias{S2TW}
\alias{S2TWP}
\alias{T2HK}
\alias{T2S}
\alias{T2TW}
\alias{TW2S}
\alias{TW2SP}
\title{The path of dictionary}
\format{character}
\usage{
S2T

S2TW

HK2S

S2HK

S2TWP

T2HK

T2S

T2TW

TW2S

TW2SP
}
\description{
The path of dictionary, and it is used by translation.
}
\keyword{datasets}

