% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hello.R
\name{converter}
\alias{converter}
\title{Converter}
\usage{
converter(type = ropencc::S2T)
}
\arguments{
\item{type}{T2S S2T}
}
\description{
create a converter. S2T means Simplified Chinese to Traditional Chinese. T2S means Traditional Chinese to Simplified Chinese.
}
\examples{
\dontrun{
cc = converter()
}
}

